Fix to work in both the APL authoring tool and Alexa devices
I was playing with this sample and realized it didn't work at all in the APL authoring tool but worked well on devices. Devices like LG and Samsung TVs use the same view host used by the APL authoring tool and that would mean this cool APL document wouldn't work on those devices. I had a developer on our team investigate and it turns out that this APL document is trying to animate the child component to display out of the parent container, which is incorrect, and should be hidden. This is what's happening in the APL authoring tool. We're updating the view host Echo devices use to behave the same way. However, as we do that I wanted to make sure this sample was updated to work in both view hosts so it doesn't break when we update the Echo devices with the fix in the future.